add card container and modify app.jsx
